requirements.txt
cloud-sql-python-connector[pg8000]
sqlalchemy
openai
asyncpg
quart
python-dotenv